# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "i9lyfe Social Media Backend API.",
        "title": "i9lyfe Backend API",
        "contact": {
            "name": "i9ine",
            "email": "oluwarinolasam@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/auth/signup/register_user": {
            "post": {
                "description": "Provide remaining user credentials",
                "consumes": [
                    "application/vnd.msgpack"
                ],
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup user - Step 3",
                "parameters": [
                    {
                        "description": "Choose a username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Choose a password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "User display name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "User birthday in milliseconds since Unix Epoch",
                        "name": "birthday",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "User bio (optional)",
                        "name": "bio",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Signup session request cookie",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signup Success",
                        "schema": {
                            "$ref": "#/definitions/signupService.signup3RespT"
                        },
                        "headers": {
                            "Set-cookie": {
                                "type": "string",
                                "description": "Authenticated user session response cookie containing auth JWT"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect or expired verification code",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        },
                        "headers": {
                            "Set-cookie": {
                                "type": "string",
                                "description": "Signup session response cookie"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        },
                        "headers": {
                            "Set-cookie": {
                                "type": "string",
                                "description": "Signup session response cookie"
                            }
                        }
                    }
                }
            }
        },
        "/auth/signup/request_new_account": {
            "post": {
                "description": "Submit email to request a new account",
                "consumes": [
                    "application/vnd.msgpack"
                ],
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup user - Step 1",
                "parameters": [
                    {
                        "description": "Provide your email address",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proceed to email verification",
                        "schema": {
                            "$ref": "#/definitions/signupService.signup1RespT"
                        },
                        "headers": {
                            "Set-cookie": {
                                "type": "string",
                                "description": "Signup session response cookie"
                            }
                        }
                    },
                    "400": {
                        "description": "An account with email already exists",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/auth/signup/verify_email": {
            "post": {
                "description": "Provide the 6-digit code sent to email",
                "consumes": [
                    "application/vnd.msgpack"
                ],
                "produces": [
                    "application/vnd.msgpack"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup user - Step 2",
                "parameters": [
                    {
                        "description": "6-digit code",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Signup session request cookie",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verified",
                        "schema": {
                            "$ref": "#/definitions/signupService.signup2RespT"
                        },
                        "headers": {
                            "Set-cookie": {
                                "type": "string",
                                "description": "Signup session response cookie"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect or expired verification code",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        },
                        "headers": {
                            "Set-cookie": {
                                "type": "string",
                                "description": "Signup session response cookie"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/appErrors.HTTPError"
                        },
                        "headers": {
                            "Set-cookie": {
                                "type": "string",
                                "description": "Signup session response cookie"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "UITypes.ClientUser": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "profilePicUrl": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "appErrors.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "signupService.signup1RespT": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "signupService.signup2RespT": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "signupService.signup3RespT": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/UITypes.ClientUser"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT API key in encrypted cookie to protect private endpoints",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}